Implement some post APIs for "order" micro service.
1. use "IdGenerator" to generate unique Id for order.
2. implement the API for comments and report.

+++ order/src/main/java/cn/edu/tongji/gohome/order/repository/CustomerCommentRepository.java
@@ -4,7 +4,8 @@ import cn.edu.tongji.gohome.order.model.CustomerCommentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface CustomerCommentRepository extends JpaRepository<CustomerCommentEntity, Integer>,
+public interface CustomerCommentRepository extends JpaRepository<CustomerCommentEntity, Long>,
         JpaSpecificationExecutor<CustomerCommentEntity> {
     CustomerCommentEntity findFirstByOrderId(long orderId);
+
 }
